Postresql and H2 compatibility fix
- partial fix for HB schema validator, UUID is bloker -> HB schema validation is disabled
- promote money values to common working SQL Type and Java type (Pg/H2)
- configuration cleanup and externalise docker profile with env variables

+++ backend/src/main/java/com/intive/shopme/model/db/DbOffer.java
@@ -24,11 +24,11 @@ public class DbOffer extends DbIdentifiable {
     @OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     private DbCategory category;
     private String baseDescription;
-    private Float basePrice;
+    private Double basePrice;
     private String extendedDescription;
-    private Float extendedPrice;
+    private Double extendedPrice;
     private String extraDescription;
-    private Float extraPrice;
+    private Double extraPrice;
     @OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     private DbOwner owner;
 }
